tp4 done
SO MANY problems with the addToCart and getQty functions - either React hook state was not being changed properly or using hard coded arrays and changed items there and returning the array was not working. if possible, ask.
also, had to create a context and learn what it was, separate files in context, provider, and hook (cart.js - useCart)